backend/requirements.txt
Django
django_polymorphic
django-viewflow
django-rest-swagger
django-filter
djangorestframework
django-rest-auth
django-seed

psycopg2-binary
python-dateutil
Markdown

pytest
pytest-django
pytest-cov
mixer # mocking data
